Revert "Update to 142.4245.2: Temprorary modify test data after IDEA update"
This reverts commit 1132652026c13c44d858e7d02976d480e522373e.

+++ idea/testData/inspections/cleanup/cleanup.kt
@@ -14,7 +14,7 @@ val x = fun foo(x: String) { }
 fun foo() {
     @loop
     for (i in 1..100) {
-        val v = oldFun2(i as Int)/* as Int*/
+        val v = oldFun2(i as Int) as Int
         /* comment */
         continue@loop
     }
